requirements.txt
vnstock
langchain
langchain-google-genai
fastapi
uvicorn
pandas
requests
python-dotenv